Use the vDSO to make syscalls
Instead of trying to clobber a random instruction in the binary, locate
a syscall instruction in our own vDSO (which is the same for all
processes) and then use that offset into the vDSO for traced processes
to perform syscalls.

It seems the old method was unreliable and didn't really work properly
unless the process was in the middle of signal handling, which is why we
had to use a true SIGSTOP (causing the parent to be notified, which is
undesirable). With this method, ptrace::attach() works as intended and
we can avoid the spurious SIGSTOP.