Set socket as blocking around send
It's technically possible for this to sendmsg with EAGAIN if the socket
queues are full. Set the socket as blocking around sendmsg to fix this.

The proper solution here would probably be to either use proper threads,
or refactor this to use async Rust so the sendmsg doesn't block other
clients... but this works for now.